no 2 A

+++ database/migrations/2025_12_07_150333_create_user_balances_table.php
@@ -9,15 +9,15 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up()
-{
-    Schema::create('user_balances', function (Blueprint $table) {
-        $table->id();
-        $table->foreignId('user_id')->constrained()->onDelete('cascade');
-        $table->bigInteger('balance')->default(0);
-        $table->timestamps();
-    });
-}
+        public function up()
+    {
+        Schema::create('user_balances', function (Blueprint $table) {
+            $table->id();
+            $table->foreignId('user_id')->constrained()->onDelete('cascade');
+            $table->decimal('balance', 12, 2)->default(0);
+            $table->timestamps();
+        });
+    }
 
 
     /**
